# Move Interface hashing and comparison to C; 2.5 to 15x speedup in micro benchmarks