… CHANGES.rst
… benchmarks/micro.py
… src/zope/interface/tests/__init__.py
… src/zope/interface/tests/test_declarations.py
… src/zope/interface/tests/test_interface.py
… src/zope/interface/tests/test_registry.py
… src/zope/interface/tests/test_sorting.py