@jamadden 2020-03-16
When running Plone's `buildout.coredev` test suite, I've had run times in the range of 22:40 to 26:11, across at least 5 different runs. With this branch, the run time was 19:06. (This is not a benchmark, and not repeatable, it's just an anecdote. But an encouraging one.)

There is one failure:
```python-traceback
File "//plone.autoform-1.8.2-py2.7.egg/plone/autoform/tests/../autoform.rst", line 428, in autoform.rst
Failed example:
    IAnotherAnonymousSchema.__module__ = 'different.module'
Exception raised:
    Traceback (most recent call last):
      File "//lib/python2.7/doctest.py", line 1315, in __run
        compileflags, 1) in test.globs
      File "<doctest autoform.rst[75]>", line 1, in <module>
        IAnotherAnonymousSchema.__module__ = 'different.module'
    TypeError: readonly attribute
```

That test is explicitly checking for an older bug:
```rest
It is possible to have interfaces/schema that have an empty __name__
attribute, specifically in some cases where a schema is dynamically
created.  In such cases, it is possible to have a subclass of
AutoExtensibleForm implement a getPrefix() function as a sufficient
condition for group naming when autoGroups is True.

    Define some unnamed schema:

    >>> class IUnknownName(Interface):
    ...     this = schema.TextLine()
    ...
    >>> IUnknownName.__name__ = ''  # dynamic schema, empty __name__

    >>> class IAnotherAnonymousSchema(Interface):
    ...     that = schema.TextLine()
    ...
    >>> IAnotherAnonymousSchema.__name__ = ''

Fix for https://github.com/zopefoundation/zope.interface/issues/31

    >>> IAnotherAnonymousSchema.__module__ = 'different.module'
```

Because equality and hashing are both based on `__name__` and `__module__`, and because various things derived from those two attributes are cached (the repr and hash code), reassigning them has long been a dangerous footgun.  Since the test is talking about dynamic schemas, I think it could be changed to actually use dynamic schemas and not run into this problem:
```python
IAnotherAnonymousSchema = InterfaceClass(
    '', 
    (Interface,) 
    {'that': schema.TextLine(), '__module__': 'different.module'})
```