@jamadden 2020-03-17
Thanks @jensens! 

I'm happy to wait for more reviews. I agree it's a big PR, plus it touches C code, so more eyes are definitely better. I'd prefer not to wait more than a couple days, though. I'd really like to get these last few PRs merged and make a release before the end of the week.